requirements.txt
web3
python-telegram-bot==13.15
requests
